Extract all test helpers into their one function files with their own tests and include in the build process.
