+++ functions/test-helpers/jsonDom.js
@@ -1,0 +1,14 @@
+/**
+ * Sample of jsonDom object containing empty nested array and objects
+ */
+const jsonDom = {
+  tagName: 'div',
+  attributes: { style: {}, className: 'column' },
+  element: null,
+  eventListeners: {},
+  parentItem: {},
+  children: [],
+  axis: 'x'
+}
+
+module.exports = jsonDom
