+++ functions/test-helpers/jsonDom.test.js
@@ -1,0 +1,12 @@
+const jsonDom = require('./jsonDom')
+
+describe('jsonDom', () => {
+  test('has sample jsonDom object', () => {
+    expect(jsonDom).toHaveProperty('attributes')
+    expect(jsonDom).toHaveProperty('children')
+    expect(jsonDom).toHaveProperty('element')
+    expect(jsonDom).toHaveProperty('eventListeners')
+    expect(jsonDom).toHaveProperty('parentItem')
+    expect(jsonDom).toHaveProperty('tagName')
+  })
+})
