+++ functions/test-helpers/nodeTree.js
@@ -1,0 +1,10 @@
+/**
+ * Sample NodeTree for testing circular references and arrays.
+ * @type {Object.<string, string|Object|Array>}
+ */
+const nodeTree = { name: 'one', parent: null, children: [] }
+nodeTree.children[0] = { name: 'child one', parent: nodeTree, children: [] }
+nodeTree.children[1] = { name: 'child two', parent: nodeTree, children: [] }
+nodeTree.children[0].children[0] = { name: 'grandchild one', parent: nodeTree.children[0], children: [] }
+
+module.exports = nodeTree
